Find location and display it on a map
This commit adds the ability to search for a location and have it displayed on a map. Nothing yet happens with this data, but it will ultimately be added to the map and table views. If the user inputs an invalid location or leaves the text field empty when tapping "Find Location" an error will be displayed on the screen.