Download StudentLocations to AppDelegate
This commit adds the StudentLocations to the AppDelegate so that is can be accessed from the list view as well as the map view. I have also moved the code for parsing through the StudentLocations and adding pins to the map to its own function to keep the code concise.