Add PostStudentLocation
Adds the ability to post the student location data to the database -- this is accomplished with the postStudentLocation API call that sends a StudentLocation object to the database.

To get the location name on the ConfirmLocationViewController I changed location: CLLocation to placemark: CLPlacemark because the CLPlacemark class contains a location propery (CLLocation) as well as a several properties that descibe the location such as name (String).