Download UserData from API and store in AppDelegate
This commit adds the function call to get user data from the API and store the response in the app delegate.

Upon a successful login the user data for the user's unique account key is downloaded and stored in the app delegate for later use.

A note about this data: this data is not authentic as the API returns randomized data for every request.