Check for existing StudentLocation and Update if needed
This commit adds the getStudentLocation function which queries the API for a student location with a matching uniqueKey. If a studentLocation is returned the StudentLocation object is assigned to a studentLocation property in the appDelegate.

This commit also adds the putStudentLocation function which is used to update a student location in the database. When creating a new student location the app will now check if a student location already exists. If a student location exists the putStudentLocation function is called instead of the postStudentLocation function.