Improve Error Handling in LoginViewController
This commit adds more robust error handling for the login procedure. This involves the creation of a OnTheMapError Type that maps a specific message to the alert view controller. To handle the mapping, a new ErrorAlert method called forStatus(_:parameter:) was created.

It is my intention to handle various other types of errors throughout the application using the OnTheMapError Type and the forStatus(_:parameter:) method. This will involve rewriting some existing code and adding additional cases to the switch statements in ErrorAlert.