Refactor ConfirmLocationViewController and add Network Activity Indicator
This commit refactors the code for ConfirmLocationViewController to make it a little more readable. Also included in this commit is the addition of an activity indicator that is displayed when there is an active network request. This provides much need visual feedback for the user after clicking on the share/finish button.