Check for and cancel existing URLSessionDataTasks
This commit moves the task variable to the global level within the class. Whenever a task is created it is assigned to this global variable. However, prior to the creation of a URLDataSessionTask the global task variable is checked and the cancel method is executed so that multiple requests are not placed.