Add shouldDownloadData Boolean
Adds a boolean to MapViewController and TableViewController that identifies whether or not the StudentLocation data should be downloaded from the API. This boolean is triggered whenever a user adds a new location to the database. This chamnge makes it so that when the user navigates back to the TableView Controller or the MapViewController after posting a location the latest data is downloaded from the server.